backend/app/models/speak.py
"""
Text-to-Speech models.
"""

from __future__ import annotations

from pydantic import BaseModel, Field


class SpeakRequest(BaseModel):
    """Request body for the TTS endpoint."""
    text: str = Field(..., min_length=1, description="Text to convert to speech")
    language: str = Field(
        default="english",
        description="Target language: 'english' or 'hindi'",
    )
