backend/requirements.txt
# ═══════════════════════════════════════════════════════════════
# IKSHAN BACKEND — Python Dependencies
# ═══════════════════════════════════════════════════════════════

# Core Framework
fastapi[standard]==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.9.1

# AI & LLM
openai==1.82.0

# HTTP Client (async)
httpx==0.28.1

# Database
supabase==2.15.2

# Environment
python-dotenv==1.1.0

# Security & Middleware
slowapi==0.1.9
python-multipart==0.0.20

# Logging
structlog==25.1.0

# Performance
orjson==3.10.18
